app/requirements.txt
python-dotenv
Flask
Flask-Mail
paypalrestsdk